Add new field "load_balance_node" to "SHOW pool_pools" and pcp_proc_info.
The new field is "1" if pgpool process is connected by a client and
the session uses the backend id as a load balance node. Users can
execute the commands to find out if there's any session that uses the
backend as the load balance node. If so, shutting down the backend may
cause session disconnection.

Discussion: https://www.pgpool.net/pipermail/pgpool-hackers/2023-July/004353.html