Test: attempt to fix 001.load_balance test timeout.
Examining the log reveals that the error was caused by bind error:
address already used.  This is strange because SO_REUSEADDR flags is
set to the listening socket.  I guess the cause is, the kernel is not
fast enough to clean up the socket used by previous pgpool process.
To fix this, existing CHECK_TIME_WAIT environment variable parameter
is modified so that it checks the socket by using netstat -tulpn,
rather than netstat -a, which is too slow.