Fix bug with pcp_proc_info.
When pcp_proc_info was invoked without "-v" option, pcp_proc_info did
not print the "statement" field.  This was due to oversight in the
frontend side of the command: forgot to add one more format siring
"%s".

This bug was in only master branch: when some new fields were added to
pcp_proc_info.

Reported-by: Bo Peng <pengbo@sraoss.co.jp>
Author: Tatsuo Ishii <ishii@postgresql.org>