Emit warning if life check has not started yet.
Currently lifecheck emits a log only when life check becomes
ready. This is inconvenient for admins since they need to keep on
watching log file until life check is ready. This commit makes the
life check process to emit additional warnings periodically until life
check has started. The interval of the warnings is wd_interval * 10
seconds. However for the first time at lifecheck starting, no warning
is emitted since it is likely that lifecheck is not ready at that
point.