Test: stabilize 039.log_backend_messages test.
In the test a query is sent to standby server right after rows are
inserted into primary server.  Due to a replication lag, the inserted
rows could not be found on the standby in slower machines.  This
commit tries to fix the issue by using synchronous replication with
remote_apply option.

Backpatch-through: v4.2