Test: renumber regression tests.
Under src/test/regression/tests we traditionally assign 001-049 for
tests for features and 050-999 for bugs. However we are running out
the range 001-049. To solve the problem, this commit renames 50 or
higher tests to 50 + 50 = 100 and higher. Also add
src/test/regression/tests/README to explain the test numbering policy.
Moreover src/test/regress.sh is modified to not misunderstand flat
files under tests is not for testing.