fix compiling on 32-bit systems
The timespec struct's tv_sec size can change between architectures.
Usually on 64 bit systems it's long, but on 32 bit systems it is frequently long long.

When the watchdog is trying to get the uptime, it is trying to get the value
as long - however on 32 bit systems this fails due to different time_t size:

| wd_json_data.c:540:66: error: passing argument 3 of 'json_get_long_value_for_key' from incompatible pointer type [-Wincompatible-pointer-types]
|   540 |         if (json_get_long_value_for_key(root, "StartupTimeSecs", &wdNode->startup_time.tv_sec))
|       |                                                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

To account for this, introduce a new helper function to get a json value as
a time_t type.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>