Explicitly stat supported minimum OpenSSL version.
That can be done in configure.ac by using
    AC_DEFINE(OPENSSL_API_COMPAT, [0x10002000L],
            [Define to the OpenSSL API version in use. This avoids deprecation warnings from newer OpenSSL versions.])

The supported minimum version is 1.0.2, which is identical to
PostgreSQL at this point.  (Actually the AC_DEFINE is brought in from
PostgreSQL).  Also this could avoid OpenSSL deprecation warnings if
your installed OpenSSL is new, like 3.0.

Moreover, add mention to installation document that required minimum
version of OpenSSL is 1.0.2.

Backpatch-through: v4.3